--- src/pages/admin.module.css
@@ -1,64 +1,0 @@
-.googleButton {
-  margin: auto 0;
-  font-size: 25px;
-  color: #13381f;
-}
-
-.signInText {
-  font-size: 20px;
-  font-weight: 600;
-  color: #13381f;
-}
-
-.header {
-  margin: auto auto;
-}
-
-.contentContainer {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 20px;
-  margin-top: 20px;
-}
-
-.container {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-}
-
-.buttonContainer {
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-  background-color: #f8b547;
-  border-radius: 10px;
-  gap: 10px;
-  width: fit-content;
-  height: 50px;
-  margin: auto 0;
-  padding: 0 15px;
-  transform: scale(1);
-  transition-timing-function: ease-in-out;
-  transition: transform 0.5s;
-}
-
-.buttonContainer:hover {
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-  background-color: #f8b547;
-  border-radius: 10px;
-  gap: 10px;
-  width: fit-content;
-  height: 50px;
-  margin: auto 0;
-  padding: 0 15px;
-  transform: scale(1.1);
-  transition-timing-function: ease-in-out;
-  transition: transform 0.5s;
-}
